requirements.txt
streamlit
openai
python-dotenv
tavily-python
llama-index-core
llama-index-llms-azure-openai
pydantic
azure-cognitiveservices-speech
requests
pillow
tenacity 